Working on lambdas
Started playing around with lambda functions. When the lambda epxression is called in the plotscript executable, the argument list is passed as the first expression in the tail and the expression to be evaluated is the second expression in the tail. I played around with that concept and begain created the lambda expression as a list. I need to figure out how to convert that into a procedure, and also how to map those arguments to the inputs for the lambda function (which need to be done prior to evaluating the expression since shadowed > captured.