Added set-property special form
Added the ability to create a map of properties in the expression class. There is a unique_ptr that points to a map where the keys and values for each property owned by the expression is held. I also wrote some tests, but will not be able to verify full coverage until get-property is implemented.