Starting work with objects in output widget
Started working on getting the graphics primitives to work. I have not included the startup file yet. I added a handleObject function that takes an expression (presumed to be an object) and its object name. This function will direct it to the right handler (line, point, text, etc.). While messing around with it, I discovered an issue that is annoying to work around - my string literal expressions are stored with the quotes. I want to update it such that the quotes only get printed along with the string literals, and are not stored as a part of them.