Moved plot implementation to expression
While implementing the continuous plot, I realized I needed access to the environment to evaluate the lambda function, so I moved the entire implementation of plots to the expression class. I kept a lot of the same code and just called the existing functions I had from within the apply function of the expression class.